Continue moving things to use-package
I think we're nearly there, now. It's really just those global modes
and hooks that need to be moved. It wouldn't really be that much work,
but I'm tired and don't really want to do that now. Maybe in a bit.

Actually, the true intent behind these changes was to eradicate
custom-set-variables and custom-set-faces, which I did—to great
effect. I'm not sure whose idea the whole "custom" system was, and I'm
sure it was added in good faith, but I really don't think it adds much
to the system. Or, perhaps, I'm just a code purist.

I'm sure these changes are fine. The editor starts without errors and
I've been able to open some files in both this here emacs-config
project as well as my work project. Like I said, these changes are
sure to be fine.